Migrate Query.project
